docs: :sparkles: using Tuples instead of Records
the previous version used Records as config option values. This made a simple config bloated because the closing bracket `}` of the Record had to be on a separate line. This version uses named Tuples which produces a much more compact config. The only downside is that there is no auto complete for Tuple but that won't be a problem since the plugin config is heavily documented with examples.

The plugin has a new config structure which is more compact and cleaner